refactor: extract JiraExportProvider+IssueCreation (#638 slice E, FINAL)
FINAL slice of #638. Closes the Jira decomposition epic.

Moved: 8 methods (export/makeURLRequest/retryContext+doc/createWithRetry/
attemptCreate/makeDescription/formattedReproductionStep/annotatedScreenshotLines)
+ 5 wire structs (JiraIssueRequest/Fields/ProjectField/IssueTypeField/Response).

Widenings:
- 3 stored props (logger/retryConfiguration/annotationRenderer): private let → let
- 5 wire structs: file-scope private → internal
- All moved methods stay private (same-file callers)

Post-slice base: init + 5 stored props only.

Cumulative on #638 (5 slices A/B/C/D/E):
- JiraExportProvider.swift: 1255 → 39 (-1216 lines, -97%)
- 5 extension files: Networking (127), ADF (114), Duplicates (197), Discovery (404), IssueCreation (390) = 1232 lines
- Net repo change: +16 lines for extension boilerplate

Closes #896.
Closes #638.
